Do not use the ICC profile for YUV conversion
Many ICC profiles do not have a direct mapping to the CICP values, so
it should be more compatible to always use the CICP values when
converting to YUV.
Removing ICC profile parsing from the encoder also means that the
Little CMS dependency is no longer needed.

Additionally, a bug with the decoder tile profile comparison has been
fixed.