Give the ICC profile box precedence over NCLX
A NCLX color information box will only be added if the image does not
have an existing ICC color profile.
This allows the ICC profile to be preserved for any image that has one.

The CCIP values that signal the color mode will be stored in the AV1
sequence header OBU.
A decoder will use the values in the AV1 sequence header OBU when the
AVIF container does not have a NCLX color information box.