Use a lookup table for the uint8_t to float conversion
ColorToYUV8 converts each color channel from [0, 255] to a floating
point value in the range of [0, 1].
Using a lookup table is more efficient than dividing each color channel
value by 255.

Because the lookup table size is known at compile-time, it is created
by a constexpr method.
This avoids having to compute it at runtime when saving an image.