Split images into tiles at most compression speeds
This can significantly reduce the memory usage when encoding a large
image, at the expense of compression efficiency.
The Fast, Medium and Slow compression speeds use a progressively larger
maximum tile size, this allows the compression efficiency to increase
as the compression speed gets slower.

Added a Very Slow speed that does not use tiles, it will ignore an
existing image grid if one exists.

Resolves #3