Fix support for writing more than 65535 items
A number of boxes wrote the itemId values as an UInt16, these boxes
will now use an UInt32 itemId if it is required.

This change also includes a performance optimization that preallocates
the ItemLocationBox and ItemInfoBox lists with the total number of items.
Every item that is saved will have ItemLocationBox and ItemInfoBox
entries.