Change the write order for the media data box items
The EXIF and/or XMP meta data is written first to improve efficiency
for readers that want to use it without reading the image data.

The alpha image data (if present) will be written before the color
image data to improve the user experience for web browsers and other
applications that may display an AVIF image as it is being streamed
over a network.