Only write a duplicated tile once
When an image has compressed tiles with identical data the data will
only be written to the file once, and the other tiles will reuse the
data written for the first tile.
This reduces the file size for a tiled image when the color and/or
alpha images have compressed tiles with identical data.