Improve the single color tile detection algorithm
The AVIF format stores the color and alpha data as separate images, so
the alpha channel value should be ignored when checking for duplicate
color tiles.
The alpha image tiles will be checked as a separate step, if the image
uses it.
This reduces false negatives when detecting tiles that have only a
single value, and allows the color and alpha images to maintain
separate duplicate tile lists.

Split the native CompressImage method into CompressColorImage and
CompressAlphaImage methods.