Optimize the BoxString Write method
The string data is now written to a temporary buffer that is either
allocated on the stack or rented from the ArrayPool.
The string buffer size includes the null-terminator that is written at
the end of the string.

Empty strings are handled as a special case where the encoding steps
are skipped and only the null-terminator is written to the file.