Vectorize the IsHomogeneous*Tile search code
BitwiseAllEqualToMasked differs from Paint.NET's BitwiseIsHomogenous in
that it requires the caller to supply a mask that selects the values to
be compared.
This feature is required because the AVIF format stores color and alpha
as separate images, so we have to mask off the unused BGRA pixel values
when checking if the tile is homogeneous.