Split the image decoding and YUV conversion steps
This will allow the managed code to allocate an output image based on
the characteristics of the AVIF image, which is a prerequisite for
converting the code to use WIC instead of a Paint.NET Surface.

Moving the image grid validation into managed code also simplifies the
native YUV conversion code.