Convert the image loading code to use WIC
This improves how high bit depth images are converted to Bgra32.
Any high bit depth images other than the supported HDR formats are
loaded as Rgba64 and converted to Bgra32 by WIC.

Currently, the only supported HDR format is the BT2020 color primaries
with the PQ transfer curve.
These images are loaded as Rgba128Float and converted to Bgra32 with
Direct2D, which provides a large improvement the image quality.